requirements.txt
fastapi
uvicorn
pydantic
uuid
python-jose[cryptography]
passlib[bcrypt]
python-multipart